+++ CHANGELOG.md
@@ -27,3 +27,10 @@ Major change on the TutorialItem model. See the documentation
 ## [1.0.2] - Updated example
 
 ## [1.0.3] - Updated default values
+
+## [1.0.4] - Allow for more control
+This update permits users to have more control over the onboarding flow
+
+- Use the `Tutorial.skipAll()` method to skip the tutorial
+- Replaced `children` property with `child`, which allows for full control over the content
+- Added the `radius` property, which will be used to set a custom radius for the `ShapeFocus.oval` shape